+++ CHANGELOG.md
@@ -4,7 +4,9 @@ x.y.z Release notes (yyyy-MM-dd)
 <!-- ### Breaking Changes - Include, if needed -->
 <!-- ### Enhancements - Include, if needed -->
 <!-- ### Fixes - Include, if needed -->
-<!-- ### Internal - Include, if needed -->
+
+### Internal
+* Add setup flag to check whether `PACECloudSDK` has been set up correctly before using it's `Kit`s
 
 ### Enhancements